Add timezone to timestamp
Before this commit, when a JsonFormatter is created with the timestamp=
True parameter, it used utcnow(). Which is not time when the log record
was created, but when the field was added. Also, utcnow is not timezone
aware, so the serialized datetime is missing the timezone designator at
the end.  This commit uses the time at which the log record was created
and sets the timezone to UTC, which results in an additional "+000" end